requirements.txt
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
jinja2>=3.1.3
pillow>=10.2.0
markitdown>=0.0.1a4
pytest>=8.0.0
